requirements.txt
absl-py==2.3.0
aiofiles==25.1.0
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
axios==0.4.0
Babel==2.9.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
cloudpathlib==0.21.1
cmake==4.0.3
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==4.2.0
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dlib==20.0.0
dnspython==2.8.0
dotenv==0.9.9
dotnet==1.5.1
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
executing==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.116.2
fasttext-wheel==0.9.2
filelock==3.18.0
filetype==1.2.0
filterpy==1.4.5
firebase_admin==7.1.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.8.0
fsspec==2025.7.0
gast==0.6.0
# Editable install with no version control (Generative-AI-Project==0.0.0)
-e b:\users\kiran\downloads2\short_answer_rag
git-filter-repo==2.47.0

greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
pytesseract
tesseract-ocr
kaggle==1.7.4.5
keras==3.10.0
kiwisolver==1.4.8
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-nvidia-ai-endpoints==1.0.0
langchain-openai==1.1.6
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.0
langcodes==3.5.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.2
magic-filter==1.0.12
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
openai==2.14.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pandas==2.3.2
parso==0.8.4
passlib==1.7.4
pillow==11.2.1
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybind11==3.0.1
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pypdf==6.5.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytweening==1.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.0
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.7.0
torchaudio==2.7.0
torchtext==0.18.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.36.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0











boto3==1.35.36
Flask==3.0.3
Flask_Cors==5.0.0
joblib==1.4.2
lightgbm==4.5.0
matplotlib==3.9.2
mlflow==2.17.0
dvc==3.53.0
dvc[s3]
nltk==3.9.1
numpy==2.1.2
pandas==2.2.3
wordcloud==1.9.3
seaborn==0.13.2